Move configs into a dedicated namespace
